#vfb-118 Fix issues with list viewer
